+++ src/main/java/com/smartx/tower/model/EveroutePackage.java
@@ -63,7 +63,7 @@ public class EveroutePackage {
 
   public static final String SERIALIZED_NAME_SIZE = "size";
   @SerializedName(SERIALIZED_NAME_SIZE)
-  private Double size;
+  private Long size;
 
   public static final String SERIALIZED_NAME_UPLOAD_TASK = "upload_task";
   @SerializedName(SERIALIZED_NAME_UPLOAD_TASK)
@@ -237,7 +237,7 @@ public class EveroutePackage {
   }
 
 
-  public EveroutePackage size(Double size) {
+  public EveroutePackage size(Long size) {
     
     this.size = size;
     return this;
@@ -250,12 +250,12 @@ public class EveroutePackage {
   @javax.annotation.Nonnull
   @ApiModelProperty(required = true, value = "")
 
-  public Double getSize() {
+  public Long getSize() {
     return size;
   }
 
 
-  public void setSize(Double size) {
+  public void setSize(Long size) {
     this.size = size;
   }
